pedidos/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
python-jose[cryptography]
requests
